\name{ji.densitygrid.ds}
\alias{ji.densitygrid.ds}
\title{Generate a density grid}
\usage{
  ji.densitygrid.ds(xvect, yvect, numints = 20)
}
\arguments{
  \item{xvect}{a numerical vector}

  \item{yvect}{a numerical vector}

  \item{numints}{a number of intervals for the grid density
  object, by default is 20}
}
\value{
  a density grid matrix
}
\description{
  Generate a density grid
}
\author{
  Gaye, A. and Isaeva, J.
}

