\name{ji.MinMax.ds}
\alias{ji.MinMax.ds}
\title{Generates a matrix with a minimum and maximum value for xvect and yvect - helper function for heatmap plot}
\usage{
  ji.MinMax.ds(xvect, yvect)
}
\arguments{
  \item{xvect}{a numerical vector}

  \item{yvect}{a numerical vector}
}
\description{
  Generates a matrix with a minimum and maximum value for
  xvect and yvect - helper function for heatmap plot
}

