\name{ag.densitygrid.ds}
\alias{ag.densitygrid.ds}
\title{Generate a density grid}
\usage{
  ag.densitygrid.ds(xvect, yvect)
}
\arguments{
  \item{xvect}{a numerical vector}

  \item{yvect}{a numerical vector}
}
\value{
  a density grid matrix
}
\description{
  Generate a density grid
}
\author{
  Gaye, A. and Isaeva, J.
}

